backend/requirements.txt
pymongo
python-dotenv
flask
flask-cors
spacy
nltk
pytesseract
pdf2image
sendgrid